implement parse JSON of HTTP /mine request
This includes a new function called find_key_index() to find a
JSON member by it's index.  The functions fails and exits
the program if it cannot find the requested key

Also clearing the buffer of global buffers and their respective
index variables before they are used in a function again
so as not to reach the end of a global array, or leave
potentially bug causing junk in an array.  This is not necessary
if I rework many or all of these global variables to non-global
in the future